+++ Shared/PlatypusAppSpec.m
@@ -387,7 +387,7 @@
     
     // Copy bundled files to Resources folder
     // .app/Contents/Resources/*
-    NSInteger numBundledFiles = [self[AppSpecKey_BundledFiles] count];
+    NSUInteger numBundledFiles = [self[AppSpecKey_BundledFiles] count];
     if (numBundledFiles) {
         [self report:@"Copying %d bundled files", numBundledFiles];
     }
@@ -800,7 +800,7 @@
     // Create bundled files string
     NSString *bundledFilesCmdString = @"";
     NSArray *bundledFiles = self[AppSpecKey_BundledFiles];
-    for (int i = 0; i < [bundledFiles count]; i++) {
+    for (NSUInteger i = 0; i < [bundledFiles count]; i++) {
         NSString *str = shortOpts ? @"-f" : @"--bundled-file";
         bundledFilesCmdString = [bundledFilesCmdString stringByAppendingString:[NSString stringWithFormat:@"%@ '%@' ", str, bundledFiles[i]]];
     }
